PIG-3617: problem with temp file deletion in MAPREDUCE operator (nezihyigitbasi via cheolsoo)
git-svn-id: https://svn.apache.org/repos/asf/pig/trunk@1556685 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/MapReduceLauncher.java
@@ -418,7 +418,10 @@ public class MapReduceLauncher extends Launcher{
         if (!"false".equalsIgnoreCase(pc.getProperties().getProperty(PigConfiguration.PIG_DELETE_TEMP_FILE))) {
             // Clean up all the intermediate data
             for (String path : intermediateVisitor.getIntermediate()) {
-                FileLocalizer.delete(path, pc);
+                // Skip non-file system paths such as hbase, see PIG-3617
+                if (Utils.hasFileSystemImpl(new Path(path), conf)) {
+                    FileLocalizer.delete(path, pc);
+                }
             }
         }
